# Conflicts:
#	.gradle/8.0/checksums/checksums.lock
#	.gradle/8.0/checksums/md5-checksums.bin
#	.gradle/8.0/checksums/sha1-checksums.bin
#	.gradle/8.0/executionHistory/executionHistory.bin
#	.gradle/8.0/executionHistory/executionHistory.lock
#	.gradle/8.0/fileHashes/fileHashes.bin
#	.gradle/8.0/fileHashes/fileHashes.lock
#	.gradle/8.0/fileHashes/resourceHashesCache.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.gradle/buildOutputCleanup/outputFiles.bin
#	.gradle/file-system.probe
#	app/build.gradle.kts
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/duancuahang/HomeShop.java
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/duancuahang/Class/ProductData.java
@@ -12,7 +12,7 @@ public class ProductData implements Serializable {
     private int quanlityProduct;
     private String descriptionProduct;
     private int sumLike;
-    private double overageCmtProduct;
+    private int overageCmtProduct;
     private String idUserProduct;
 
     public ProductData() {
@@ -98,15 +98,15 @@ public class ProductData implements Serializable {
         this.sumLike = sumLike;
     }
 
-    public double getOverageCmtProduct() {
+    public int getOverageCmtProduct() {
         return overageCmtProduct;
     }
 
-    public void setOverageCmtProduct(double overageCmtProduct) {
+    public void setOverageCmtProduct(int overageCmtProduct) {
         this.overageCmtProduct = overageCmtProduct;
     }
 
-    public ProductData(String idProduct, String nameProduct,  int priceProduct, String keyCategoryProduct, String keyManufaceProduct, int quanlityProduct, String descriptionProduct, int sumLike, double overageCmtProduct, String idUserProduct) {
+    public ProductData(String idProduct, String nameProduct,  int priceProduct, String keyCategoryProduct, String keyManufaceProduct, int quanlityProduct, String descriptionProduct, int sumLike, int overageCmtProduct, String idUserProduct) {
         this.idProduct = idProduct;
         this.nameProduct = nameProduct;
 //        this.urlImageProduct = urlImageProduct;
